metodo de listado  inscipciones
este metodo depende de los metodos de buscar alumno y buscar materia cuando esten subidos sus cambias los acoplo a este metodo para que funcione , mientras tiene lineas comentadas

+++ src/universidadAPP/Entidades/Inscripcion.java
@@ -11,20 +11,20 @@ package universidadAPP.Entidades;
  */
 public class Inscripcion {
     private int idInscripto;
-    private int nota;
+    private double nota;
     private Ealumno alumno;
     private Materia materia;
 
     public Inscripcion() {
     }
 
-    public Inscripcion(int nota, Ealumno alumno, Materia materia) {
+    public Inscripcion(double nota, Ealumno alumno, Materia materia) {
         this.nota = nota;
         this.alumno = alumno;
         this.materia = materia;
     }
 
-    public Inscripcion(int idInscripto, int nota, Ealumno alumno, Materia materia) {
+    public Inscripcion(int idInscripto, double nota, Ealumno alumno, Materia materia) {
         this.idInscripto = idInscripto;
         this.nota = nota;
         this.alumno = alumno;
@@ -39,11 +39,11 @@ public class Inscripcion {
         this.idInscripto = idInscripto;
     }
 
-    public int getNota() {
+    public double getNota() {
         return nota;
     }
 
-    public void setNota(int nota) {
+    public void setNota(double nota) {
         this.nota = nota;
     }
